refactor(server): 提取 ServerError 到新的模块
